- Las reseñas se le hacen a un parking y rating ya sale.

+++ app/src/main/java/com/example/parkingcompose/model/Parking.kt
@@ -12,10 +12,10 @@ data class Parking(
     val name: String = "",
     val description: String = "",
     val image: String = "", // URL del Storage
-    val parkingRating: Float = 0.0f,
+    val parkingRating: Double = 0.0,
     val reviewList: List<Review> = emptyList(),
     val tagList: List<Tag> = emptyList(),
     val priceMinute: Float = 0.0f,
     val checked: Boolean = false,
-    val createdBy : String = ""
+    val createdBy: String = ""
 )
